[FEAT] Do tests directly with flake8
Instead of just calling the class or `main` it should be also run with `flake8`
in case it doesn't use the class as expected. To do that it either checks if
it's installed using `pip` in the right location. Otherwise it installs it if
`TEST_FLAKE8_INSTALL` is set to `1`. Afterwards it uninstalls it, if it wasn't
already installed.