Make Feature test functional
For each `Feature` it knows but has no `__future__` import, it is checking that
the `Feature` will be introduced in a later version. But creating that test was
incomplete, as it used the wrong looping variable, as well as not returning the
actual test function.

It now also just does a standard `assert` to verify that
`__future__.all_feature_names` is correctly defined.