Fix checking imports against minimum version (#7)
The minimum version should only reduce the number of missing imports, by not
including imports which are not available yet and by those which are already
mandatory. This fixed two issues:

One issue was that it would not issue a warning if an import was used which
wasn't already available on that version yet. It should only suppress a warning
when the import wasn't imported, as it isn't available yet.

The other issue was that it would still issue a warning if the feature was made
mandatory in exactly the same version that is set to the minimum. But in that
case it should suppress the warning, when it is missing, because that feature
is already mandatory. (This was the actual issue reported in #7)